+++ bcs-services/bcs-cluster-manager/internal/actions/namespace/create.go
@@ -77,7 +77,7 @@ func (ca *CreateAction) createNamespace() error {
 	return ca.model.CreateNamespace(ca.ctx, newNs)
 }
 
-func (ca *CreateAction) setResp(code uint64, msg string) {
+func (ca *CreateAction) setResp(code uint32, msg string) {
 	ca.resp.Code = code
 	ca.resp.Message = msg
 	ca.resp.Result = (code == types.BcsErrClusterManagerSuccess)
